[Refactor] ssd.py, ssd_file_manager.py 파일 분리
